app/requirements.txt
fastapi~=0.85.1
google-cloud-bigquery==3.3.5
uvicorn==0.18.3
redis==4.3.4
pytest==7.1.3